Namespace all the steampress stuff on application
