+++ Tests/SteamPressTests/ProviderTests.swift
@@ -8,17 +8,17 @@ class ProviderTests: XCTestCase {
         app.middleware.use(BlogRememberMeMiddleware())
         app.middleware.use(SessionsMiddleware(session: app.sessions.driver))
 
-        app.blogRepositories.use { application in
+        app.steampress.blogRepositories.use { application in
             return InMemoryRepository(eventLoop: application.eventLoopGroup.next())
         }
         
-        let numberGenerator = app.randomNumberGenerators.generator
+        let numberGenerator = app.steampress.randomNumberGenerators.generator
         XCTAssertTrue(type(of: numberGenerator) == RealRandomNumberGenerator.self)
         
-        let blogPresenter = app.blogPresenters.blogPresenter
+        let blogPresenter = app.steampress.blogPresenters.blogPresenter
         XCTAssertTrue(type(of: blogPresenter) == ViewBlogPresenter.self)
         
-        let blogAdminPresenter = app.adminPresenters.adminPresenter
+        let blogAdminPresenter = app.steampress.adminPresenters.adminPresenter
         XCTAssertTrue(type(of: blogAdminPresenter) == ViewBlogAdminPresenter.self)
         
         app.shutdown()
